fix: workflow approval, refactor
