+++ packages/plugin-workflow-approval/src/client/user-interface/pc/block/apply-list/FeatureModal.view.tsx
@@ -8,14 +8,14 @@ import {
 } from '@tachybase/client';
 import { ProviderContextWorkflow } from '@tachybase/module-workflow/client';
 
-import { ActionBarProvider } from '../table-initiations/apply-button/ActionBar.provider';
-import { ApplyActionStatusProvider } from '../table-initiations/apply-button/ActionStatus.provider';
-import { WithdrawActionProvider } from '../table-initiations/apply-button/ActionWithdraw.provider';
-import { useSubmitCreate } from '../table-initiations/apply-button/hooks/useSubmitCreate';
-import { useActionReminder } from '../table-initiations/hooks/useActionReminder';
-import { useActionResubmit } from '../table-initiations/hooks/useActionResubmit';
-import { useWithdrawAction } from '../table-initiations/hooks/useWithdrawAction';
-import { ProviderActionReminder } from '../table-initiations/providers/ActionReminder.provider';
+import { useActionReminder } from '../common/hooks/useActionReminder';
+import { useActionResubmit } from '../common/hooks/useActionResubmit';
+import { useSubmitCreate } from '../common/hooks/useSubmitCreate';
+import { useWithdrawAction } from '../common/hooks/useWithdrawAction';
+import { ActionBarProvider } from '../common/providers/ActionBar.provider';
+import { ProviderActionReminder } from '../common/providers/ActionReminder.provider';
+import { ApplyActionStatusProvider } from '../common/providers/ActionStatus.provider';
+import { WithdrawActionProvider } from '../common/providers/ActionWithdraw.provider';
 import { getSchemaFeatureModal } from './FeatureModal.schema';
 
 export const ViewFeatureModal = (props) => {
