+++ packages/plugin-workflow-approval/src/client/user-interface/pc/block/common/hooks/useSubmitCreate.tsx
@@ -10,7 +10,7 @@ import { useField, useForm } from '@tachybase/schema';
 
 import _ from 'lodash';
 
-import { useContextApprovalStatus } from '../ActionStatus.provider';
+import { useContextApprovalStatus } from '../providers/ActionStatus.provider';
 
 export function useSubmitCreate() {
   const form = useForm();
