+++ packages/plugin-workflow-approval/src/client/user-interface/pc/block/table-initiations/CheckContent.view.tsx
@@ -10,19 +10,19 @@ import { useForm } from '@tachybase/schema';
 
 import { FormBlockProvider } from '../../../../common/components/FormBlock.provider';
 import { ContextWithActionEnabled } from '../../common/WithActionEnabled.provider';
-import { SchemaComponentContextProvider } from '../common/SchemaComponent.provider';
+import { useActionReminder } from '../common/hooks/useActionReminder';
+import { useActionResubmit } from '../common/hooks/useActionResubmit';
+import { useWithdrawAction } from '../common/hooks/useWithdrawAction';
+import { ProviderActionReminder } from '../common/providers/ActionReminder.provider';
+import { SchemaComponentContextProvider } from '../common/providers/SchemaComponentContextProvider';
+import { WithdrawActionProvider } from '../common/providers/WithdrawAction.provider';
 import { getSchemaCheckContent } from './CheckContent.schema';
-import { useActionReminder } from './hooks/useActionReminder';
-import { useActionResubmit } from './hooks/useActionResubmit';
 import { useDestroyAction } from './hooks/useDestroyAction';
 import { useFormBlockProps } from './hooks/useFormBlockProps';
 import { useSubmitUpdate } from './hooks/useSubmitUpdate';
-import { useWithdrawAction } from './hooks/useWithdrawAction';
 import { ActionBarProvider } from './providers/ActionBar.provider';
-import { ProviderActionReminder } from './providers/ActionReminder.provider';
 import { ProviderActionResubmit } from './providers/ActionResubmit.provider';
 import { ProviderApplyActionStatus } from './providers/ApplyActionStatus.provider';
-import { WithdrawActionProvider } from './WithdrawAction.provider';
 
 export const ViewCheckContent = (props) => {
   const { approval, workflow } = props;
