--- packages/plugin-workflow-approval/src/client/user-interface/pc/block/table-initiations/apply-button/hooks/useActionResubmit.ts
@@ -1,5 +1,0 @@
-export function useActionResubmit() {
-  return {
-    run() {},
-  };
-}
