--- packages/plugin-workflow-approval/src/client/user-interface/pc/block/table-initiations/apply-button/hooks/useWithdrawAction.tsx
@@ -1,3 +1,0 @@
-export function useWithdrawAction() {
-  return { run() {} };
-}
